+++ Kavita.Models/DTOs/Scrobbling/PlusSeriesDto.cs
@@ -10,7 +10,7 @@ public class PlusSeriesRequestDto
     public long? MalId { get; set; }
     public string? GoogleBooksId { get; set; }
     public string? MangaDexId { get; set; }
-    public long? MangabakaId { get; set; }
+    public int? MangabakaId { get; set; }
     /// <summary>
     /// ComicBookRoundup Id
     /// </summary>
